code/get_pulls/requirements.txt
requests==2.18.4
python-dateutil==2.6.1
